requirements-test.txt
-r requirements.txt
-r requirements-build.txt

# examples
django-environ==0.4.5
django-filter==2.2.0
django-import-export==2.0.2
django-storages==1.9.1
Pillow==7.1.1
django-ckeditor==5.9.0
djangocms-text-ckeditor==3.9.0
djangocms-snippet==2.3.0
djangocms-style==2.3.0
psycopg2==2.8.5

# test
pytest==5.4.1
pytest-cov==2.8.1
pytest-django==3.9.0
black==19.10b0
flake8==3.7.9
selenium==3.141.0
